+++ src/main/kotlin/veryFit2googleFit/app/AppCommand.kt
@@ -2,8 +2,8 @@ package veryFit2googleFit.app
 
 import com.github.ajalt.clikt.core.CliktCommand
 
-object AppCommand: CliktCommand() {
-    override fun run() {
-        TODO("Not yet implemented")
-    }
+object AppCommand : CliktCommand() {
+	override fun run() {
+		TODO("Not yet implemented")
+	}
 }
