+++ src/main/kotlin/veryFit2googleFit/vfit/VFitFiles.kt
@@ -1,0 +1,11 @@
+package veryFit2googleFit.vfit
+
+import java.io.File
+import java.nio.file.Path
+
+class VFitFiles(root: Path) {
+	val syncDirPath = root.resolve("sync")
+	val syncDir = syncDirPath.toFile()
+	val syncLogFiles: Iterable<File>
+		get() = syncDir.listFiles { file -> file.extension == "log" }?.asIterable() ?: emptyList()
+}
