+++ src/test/kotlin/veryFit2googleFit/AppTest.kt
@@ -7,8 +7,9 @@ import kotlin.test.Test
 import kotlin.test.assertNotNull
 
 class AppTest {
-    @Test fun testAppHasAGreeting() {
-        val classUnderTest = App()
-        assertNotNull(classUnderTest.greeting, "app should have a greeting")
-    }
+	@Test
+	fun testAppHasAGreeting() {
+		val classUnderTest = App()
+		assertNotNull(classUnderTest.greeting, "app should have a greeting")
+	}
 }
